+++ TOSCAGeneratorClient/src/main/java/org/acumos/designstudio/toscagenerator/service/ProtobufGeneratorService.java
@@ -1,22 +1,22 @@
 /*-
- * ===============LICENSE_START=======================================================
- * Acumos
- * ===================================================================================
- * Copyright (C) 2017 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
- * ===================================================================================
- * This Acumos software file is distributed by AT&T and Tech Mahindra
- * under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *  
- *      http://www.apache.org/licenses/LICENSE-2.0
- *  
- * This file is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * ===============LICENSE_END=========================================================
- */
+* ===============LICENSE_START=======================================================
+* Acumos
+* ===================================================================================
+* Copyright (C) 2017 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+* ===================================================================================
+* This Acumos software file is distributed by AT&T and Tech Mahindra
+* under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*  
+*      http://www.apache.org/licenses/LICENSE-2.0
+*  
+* This file is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+* ===============LICENSE_END=========================================================
+*/
 
 package org.acumos.designstudio.toscagenerator.service;
 
@@ -41,6 +41,7 @@ import org.acumos.designstudio.toscagenerator.vo.protobuf.Operation;
 import org.acumos.designstudio.toscagenerator.vo.protobuf.Option;
 import org.acumos.designstudio.toscagenerator.vo.protobuf.OutputMessage;
 import org.acumos.designstudio.toscagenerator.vo.protobuf.ProtoBufClass;
+import org.acumos.designstudio.toscagenerator.vo.protobuf.Service;
 import org.acumos.designstudio.toscagenerator.vo.protobuf.SortComparator;
 import org.acumos.designstudio.toscagenerator.vo.protobuf.SortFactory;
 import org.slf4j.Logger;
@@ -51,10 +52,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.google.gson.Gson;
 
-/**
- * @author AB00343130
- */
-
 public class ProtobufGeneratorService {
 
 	boolean isMessage = false;
@@ -63,7 +60,6 @@ public class ProtobufGeneratorService {
 	int servicesLineCount = 0;
 
 	private static final Logger logger = LoggerFactory.getLogger(ProtobufGeneratorService.class);
-	// Protobuf's variable declaration section started
 	final Pattern pattern = Pattern.compile("\\((.*?)\\)");
 	private List<MessageargumentList> messageargumentList = null;
 	private List<InputMessage> listOfInputMessages = null;
@@ -79,22 +75,22 @@ public class ProtobufGeneratorService {
 	private List<String> listOfInputAndOutputMessage = null;
 	private List<Option> listOfOption = null;
 
-	// Start to declare new variables as per new complex protobuf file.
-	private ProtoBufClass duplicateProtoBufClass = null;
 
 	/**
-	 * 
+	 * This method create the json format of .proto file.
 	 * @param solutionId
 	 *            solution ID
 	 * @param version
 	 *            version
 	 * @param localMetadataFile
-	 * metadata file
+	 *            metadata file
 	 * @return Protobuf JSON
-	 * @throws ServiceException On failure
+	 *             string of protobuf json
+	 * @throws ServiceException
+	 *             On failure
 	 */
 	public String createProtoJson(String solutionId, String version, File localMetadataFile) throws ServiceException {
-		logger.debug("-------------- CreateProtoJson() started ---------------");
+		logger.debug("CreateProtoJson() started");
 		protoBufClass = new ProtoBufClass();
 		messageBodyList = new ArrayList<>();
 		listOfInputAndOutputMessage = new ArrayList<>();
@@ -112,30 +108,28 @@ public class ProtobufGeneratorService {
 			Gson gson = new Gson();
 			protoBufToJsonString = gson.toJson(protoBufClass);
 			try {
-				List<MessageBody> expendedmessageBodyList = constructSubMessageBody(protoBufToJsonString);
+				List<MessageBody> expendedmessageBodyList = constructListOfMessages(protoBufToJsonString);
 				protoBufClass.setListOfMessages(expendedmessageBodyList);
 				Gson gson1 = new Gson();
 				protoBufToJsonString = gson1.toJson(protoBufClass);
-				// System.out.println("protoBufToJsonString----------------" +
-				// protoBufToJsonString);
 			} catch (Exception ex) {
-				logger.error(" --------------- Exception Occured  constructSubMessageBody() ---------", ex);
+				logger.error("Exception Occured  constructListOfMessages()", ex);
 				ex.printStackTrace();
 			}
 			isMessage = false;
 			isItservice = false;
 			isItMessage = false;
 			servicesLineCount = 0;
-			logger.debug("-------------- CreateProtoJson() end ---------------");
+			logger.debug("CreateProtoJson() end");
 		} catch (ServiceException ex) {
 			logger.error(
-					" --------------- Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json--------------",
+					"Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json",
 					ex);
 			throw new ServiceException(ex.getMessage(), ex.getErrorCode(), ex.getErrorDesc(), ex.getCause());
 
 		} catch (Exception ex) {
 			logger.error(
-					" --------------- Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json--------------",
+					"Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json",
 					ex);
 
 		} finally {
@@ -146,18 +140,13 @@ public class ProtobufGeneratorService {
 					fr.close();
 			} catch (IOException ex) {
 				logger.error(
-						" --------------- Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json--------------",
+						"Exception Occured  CreateProtoJson() when Reading the protobuf file and generating protobuf json",
 						ex);
 			}
 		}
 		return protoBufToJsonString;
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @throws ServiceException
-	 */
 	private void parseLine(String line) throws ServiceException {
 		try {
 			// construct syntax
@@ -177,7 +166,7 @@ public class ProtobufGeneratorService {
 			// start package
 
 			if (line.startsWith("message")) {
-				logger.debug("-------------- costructMessage() strated ---------------");
+				logger.debug("costructMessage() strated");
 				isMessage = false;
 				messageBody = new MessageBody();
 
@@ -190,11 +179,11 @@ public class ProtobufGeneratorService {
 				messageBodyList.add(messageBody);
 				protoBufClass.setListOfMessages(messageBodyList);
 				isMessage = false;
-				logger.debug("-------------- costructMessage() end ---------------");
+				logger.debug("costructMessage() end");
 			}
 
 			if (line.startsWith("service")) {
-				logger.debug("-------------- constructService() started ---------------");
+				logger.debug("constructService() started");
 				service = new org.acumos.designstudio.toscagenerator.vo.protobuf.Service();
 				service = constructService(line, service);
 
@@ -229,23 +218,18 @@ public class ProtobufGeneratorService {
 				}
 			} else if (isItservice && line.contains("}") && !line.isEmpty()) {
 				isItservice = false;
-				logger.debug("-------------- constructService() end ---------------");
+				logger.debug("constructService() end");
 			}
 		} catch (Exception ex) {
-			logger.error(" --------------- Exception Occured  parseLine() --------------", ex);
-			throw new ServiceException(" --------------- Exception Occured while parsing protobuf --------------",
+			logger.error("Exception Occured  parseLine()", ex);
+			throw new ServiceException("Exception Occured while parsing protobuf",
 					Properties.getDecryptionErrorCode(), "Error while parsing protoBuf file", ex.getCause());
 		}
 
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @return
-	 */
 	private String constructSyntax(String line) {
-		logger.debug("-------------- constructSyntax() strated ---------------");
+		logger.debug("constructSyntax() strated");
 		String removequotes = "";
 		try {
 			if (line.startsWith("syntax")) {
@@ -255,31 +239,20 @@ public class ProtobufGeneratorService {
 				removequotes = trimString.replaceAll("^\"|\"$", "");
 			}
 		} catch (Exception ex) {
-			logger.error(" --------------- Exception Occured  constructSyntax() --------------", ex);
+			logger.error("Exception Occured  constructSyntax()", ex);
 		}
-		logger.debug("-------------- constructSyntax() end ---------------");
+		logger.debug("constructSyntax() end");
 		return removequotes;
 
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @return
-	 */
 	private String constructPackage(String line) {
-		logger.debug("-------------- constructPackage() strated ---------------");
+		logger.debug("constructPackage() strated");
 		String[] fields = line.split(" ");
-		logger.debug("-------------- constructPackage() end ---------------");
+		logger.debug("constructPackage() end");
 		return fields[1].replace(";", "");
-
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @return
-	 */
 	private String constructOption(String line) {
 		Option optionInstance = new Option();
 		StringTokenizer st = new StringTokenizer(line, " ");
@@ -297,24 +270,10 @@ public class ProtobufGeneratorService {
 		return "";
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @param messageBody
-	 * @param messageargumentList
-	 * @return
-	 */
 	private MessageBody costructMessage(String line, MessageBody messageBody,
 			List<MessageargumentList> messageargumentList) {
 		try {
 			if (line.startsWith("message")) {
-				/*
-				 * int openCurlybacketPosition; String messageValue = "";
-				 */
-				/*
-				 * openCurlybacketPosition = line.indexOf("{"); messageValue = line.substring(8,
-				 * openCurlybacketPosition); messageBody.setMessageName(messageValue.trim());
-				 */
 				String[] fields = line.split(" ");
 				messageBody.setMessageName(fields[1]);
 			}
@@ -327,8 +286,7 @@ public class ProtobufGeneratorService {
 
 					for (int i = 0; i < fields.length; i++) {
 						if (!fields[i].isEmpty()) {
-							if (st.countTokens() == 5) {// count total number of
-														// token in line.
+							if (st.countTokens() == 5) {// count total number of token in line.
 								messageargumentListInstance.setRole(fields[i]);
 								messageargumentListInstance.setType(fields[i + 1]);
 								messageargumentListInstance.setName(fields[i + 2]);
@@ -358,27 +316,14 @@ public class ProtobufGeneratorService {
 				isMessage = true;
 			}
 		} catch (Exception ex) {
-			logger.error(" --------------- Exception Occured  costructMessage() --------------", ex);
+			logger.error("Exception Occured  costructMessage()", ex);
 		}
 		return messageBody;
 	}
 
-	/**
-	 * 
-	 * @param line
-	 * @param service
-	 * @return
-	 */
 	private org.acumos.designstudio.toscagenerator.vo.protobuf.Service constructService(String line,
 			org.acumos.designstudio.toscagenerator.vo.protobuf.Service service) {
 		try {
-			/*
-			 * String servicesName = ""; int openCurlybacketPosition;
-			 */
-			/*
-			 * openCurlybacketPosition = line.indexOf("{"); servicesName = line.substring(8,
-			 * openCurlybacketPosition); service.setName(servicesName.trim());
-			 */
 			String[] fields = line.split(" ");
 			service.setName(fields[1]);
 			if (line.contains("}")) {
@@ -387,19 +332,14 @@ public class ProtobufGeneratorService {
 				isItservice = true;
 			}
 		} catch (Exception ex) {
-			logger.error(" --------------- Exception Occured  constructService() --------------", ex);
+			logger.error("Exception Occured  constructService()", ex);
 		}
 		return service;
 
 	}
 
-	/**
-	 * 
-	 * @param inputParameterString
-	 * @return
-	 */
 	private List<InputMessage> constructInputMessage(String inputParameterString) {
-		logger.debug("-------------- constructInputMessage() strated ---------------");
+		logger.debug("constructInputMessage() strated");
 		listOfInputMessages = new ArrayList<InputMessage>();
 		try {
 
@@ -411,20 +351,15 @@ public class ProtobufGeneratorService {
 				listOfInputAndOutputMessage.add(inputMessage.getInputMessageName());
 			}
 		} catch (Exception ex) {
-			logger.error(" --------------- Exception Occured  constructInputMessage() --------------", ex);
+			logger.error("Exception Occured  constructInputMessage()", ex);
 		}
-		logger.debug("-------------- constructInputMessage() end ---------------");
+		logger.debug("constructInputMessage() end");
 		return listOfInputMessages;
 
 	}
 
-	/**
-	 * 
-	 * @param outPutParameterString
-	 * @return
-	 */
 	private List<OutputMessage> constructOutputMessage(String outPutParameterString) {
-		logger.debug("-------------- constructOutputMessage() strated ---------------");
+		logger.debug("constructOutputMessage() strated");
 		listOfOputPutMessages = new ArrayList<OutputMessage>();
 		try {
 			String[] outPutParameterArray = outPutParameterString.split(",");
@@ -435,257 +370,116 @@ public class ProtobufGeneratorService {
 				listOfInputAndOutputMessage.add(outputMessage.getOutPutMessageName());
 			}
 		} catch (Exception ex) {
-			logger.error("-------------- constructOutputMessage() end ---------------", ex);
+			logger.error("constructOutputMessage() end", ex);
 		}
-		logger.debug("-------------- constructOutputMessage() end ---------------");
+		logger.debug("constructOutputMessage() end");
 		return listOfOputPutMessages;
 	}
 
-	/**
-	 * 
-	 * @param protoBufToJsonString
-	 * @return
-	 */
-	private List<MessageBody> constructSubMessageBody(String protoBufToJsonString) throws Exception {
-		logger.debug("-------------- constructconstructSubMessageBody() strated ---------------");
-		List<MessageBody> expendedmessageBodyList = new ArrayList<MessageBody>();
-		try {
-			// private List<MessageBody> duplicateexpendedmessageBodyList = new
-			// ArrayList<MessageBody>();
-			List<MessageBody> parentNotExistList = new ArrayList<MessageBody>();
-			List<String> expendedChildmessageBodyList = new ArrayList<String>();
-			ObjectMapper mapper = new ObjectMapper();
-			mapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
-			mapper.configure(DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT, true);
-			protoBufClass = mapper.readValue(protoBufToJsonString, ProtoBufClass.class);
-			duplicateProtoBufClass = mapper.readValue(protoBufToJsonString, ProtoBufClass.class);
-			List<MessageBody> sourceMessageBody = protoBufClass.getListOfMessages();
-			List<MessageBody> destinationMessageBody = duplicateProtoBufClass.getListOfMessages();
-			// String[] basicProtobufTypes = { "string", "int32", "int64" };
-			String basicProtobufTypes = Properties.getProtobufBasicType();
-			String[] basicProtobufTypesArray = basicProtobufTypes.split(",");
-			List<String> basicProtobufTypesList = Arrays.asList(basicProtobufTypesArray);
-			int i;
-			boolean isSubMessageFound = false;
-			for (i = 0; i < sourceMessageBody.size(); i++) {
-				String sourceMessageName = sourceMessageBody.get(i).getMessageName();
-				List<MessageargumentList> messageargumentList = sourceMessageBody.get(i).getMessageargumentList();
-				List<MessageargumentList> comlpexmessageargumentList = new ArrayList<MessageargumentList>();
-				int j = 0;
-				for (j = 0; j < destinationMessageBody.size(); j++) {
-					String destinationMessageName = destinationMessageBody.get(j).getMessageName(); // change
-																									// before
-																									// it
-																									// was
-																									// messageBody
-					if (sourceMessageName.equals(destinationMessageName)) {
-						continue;
-					}
-					List<MessageargumentList> parentmessageNamemessageargumentList = destinationMessageBody.get(j)
-							.getMessageargumentList();
-					if (null != parentmessageNamemessageargumentList
-							&& !parentmessageNamemessageargumentList.isEmpty()) {
-						for (int k = 0; k < parentmessageNamemessageargumentList.size(); k++) {
-							String type = parentmessageNamemessageargumentList.get(k).getType();
-							if (!basicProtobufTypesList.contains(type) && type.equals(sourceMessageName)) {
-								isSubMessageFound = true;
-								MessageBody expendedmessageBody = null;
-								List<MessageargumentList> messageargumentListInstanceList = null;
-								String message = "";
-								if (expendedmessageBodyList != null && !expendedmessageBodyList.isEmpty()) {
-									for (int ei = 0; ei < expendedmessageBodyList.size(); ei++) {
-										expendedmessageBody = expendedmessageBodyList.get(ei);
-										message = expendedmessageBody.getMessageName();
-										if (message == destinationMessageBody.get(j).getMessageName()) {
-											messageargumentListInstanceList = expendedmessageBody
-													.getMessageargumentList();
-											break;
-										} else {
-											messageargumentListInstanceList = new ArrayList<MessageargumentList>();
-											expendedmessageBody = new MessageBody();
-										}
+	private List<MessageBody> constructListOfMessages(String protoBufToJsonString) throws Exception {
 
-									}
-								} else {
-									messageargumentListInstanceList = new ArrayList<MessageargumentList>();
-									expendedmessageBody = new MessageBody();
-								}
-								if (expendedChildmessageBodyList.contains(destinationMessageName)) {
-									MessageargumentList messageargumentListInstance = new MessageargumentList();
-									ComplexType complexType = null;
-									for (MessageBody messageBody0 : expendedmessageBodyList) {
-										List<MessageargumentList> list = messageBody0.getMessageargumentList();
-										MessageargumentList messageArgumentList = null;
-										int messageArgumentListIndex = 0;
-										for (int jj = 0; jj < list.size(); jj++) {
-											messageArgumentList = messageBody0.getMessageargumentList().get(jj);
-											if (messageArgumentList.getComplexType() != null) {
-												complexType = messageArgumentList.getComplexType();
-												messageArgumentListIndex = jj;
-											}
-										}
-										// MessageargumentList
-										// messageArgumentList =
-										// messageBody0.getMessageargumentList().get(0);
+		List<MessageBody> listOfMessages = new ArrayList<MessageBody>();
+		MessageBody messageBody = null;
+		ObjectMapper mapper = new ObjectMapper();
+		mapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
+		mapper.configure(DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT, true);
+		protoBufClass = mapper.readValue(protoBufToJsonString, ProtoBufClass.class);
 
-										// complexType =
-										// messageArgumentList.getComplexType();
-										List<MessageargumentList> complexMessageargumentList = complexType
-												.getMessageargumentList();
-										ComplexType complexTypeObject = null;
-										for (int ii = 0; ii < complexMessageargumentList.size(); ii++) {
-											MessageargumentList MessageargumentList = complexMessageargumentList
-													.get(ii);
-											if (MessageargumentList.getType().equals(sourceMessageName)) {
-												String parentTag = MessageargumentList.getTag();
-												List<MessageargumentList> comlpexmessageargumentList20 = new ArrayList<MessageargumentList>();
-												complexTypeObject = new ComplexType();
-												for (int l = 0; l < messageargumentList.size(); l++) {
-													MessageargumentList messageargumentList12 = (MessageargumentList) messageargumentList
-															.get(l);
-													MessageargumentList complexMessageargumentInstance = new MessageargumentList();
-													complexMessageargumentInstance
-															.setRole(messageargumentList12.getRole());
-													complexMessageargumentInstance
-															.setType(messageargumentList12.getType());
-													complexMessageargumentInstance
-															.setName(messageargumentList12.getName());
-													complexMessageargumentInstance
-															.setTag(parentTag + "." + messageargumentList12.getTag());
-													comlpexmessageargumentList20.add(complexMessageargumentInstance);
-												}
-												complexTypeObject.setMessageName(sourceMessageName);
-												complexTypeObject.setMessageargumentList(comlpexmessageargumentList20);
-												messageargumentListInstance.setRole(MessageargumentList.getRole());
-												messageargumentListInstance.setType(MessageargumentList.getType());
-												messageargumentListInstance.setComplexType(complexTypeObject);
-												messageargumentListInstance.setName(MessageargumentList.getName());
-												messageargumentListInstance.setTag(MessageargumentList.getTag());
-												complexMessageargumentList.set(ii, messageargumentListInstance);
-												complexType.setMessageargumentList(complexMessageargumentList);
-												messageArgumentList.setComplexType(complexType);
-												messageBody0.getMessageargumentList().set(messageArgumentListIndex,
-														messageArgumentList);
+		Service service = protoBufClass.getService();
+		List<Operation> listOfOperations = service.getListOfOperations();
+		List<InputMessage> inputMessages = null;
+		String inputMessageName = null;
 
-											}
+		List<OutputMessage> outputMessages = null;
+		String outputMessageName = null;
+		boolean msgExist = false;
+		for (Operation operation : listOfOperations) {
+			inputMessages = operation.getListOfInputMessages();
+			for (InputMessage inputmsg : inputMessages) {
+				inputMessageName = inputmsg.getInputMessageName();
+				// expanded
+				messageBody = constructExpandedMessageBody(inputMessageName, protoBufClass.getListOfMessages(), "");
+				if (null != messageBody) {
+					listOfMessages.add(messageBody);
+				}
 
-										}
+			}
+			outputMessages = operation.getListOfOutputMessages();
 
-									}
-
-								} else {
-									expendedmessageBody.setMessageName(destinationMessageBody.get(j).getMessageName());
-									MessageargumentList messageargumentListInstance = new MessageargumentList();
-									MessageargumentList messageargumentListInstanceSympletype = null;
-									List<MessageargumentList> messageargumentListInstanceSympletypePreList = new ArrayList<MessageargumentList>();
-									List<MessageargumentList> messageargumentListInstanceSympletypePostList = new ArrayList<MessageargumentList>();
-									int simpleTypePreIndex = k;
-									int simpleTypePostIndex = k + 1;
-									if (simpleTypePreIndex > 0) {
-										for (int kk = 0; kk < simpleTypePreIndex; kk++) {
-											messageargumentListInstanceSympletype = new MessageargumentList();
-											messageargumentListInstanceSympletype
-													.setRole(parentmessageNamemessageargumentList.get(kk).getRole());
-											messageargumentListInstanceSympletype
-													.setType(parentmessageNamemessageargumentList.get(kk).getType());
-											messageargumentListInstanceSympletype
-													.setName(parentmessageNamemessageargumentList.get(kk).getName());
-											messageargumentListInstanceSympletype
-													.setTag(parentmessageNamemessageargumentList.get(kk).getTag());
-											messageargumentListInstanceSympletypePreList
-													.add(messageargumentListInstanceSympletype);
-										}
-
-									}
-									List<MessageargumentList> postList = parentmessageNamemessageargumentList
-											.subList(simpleTypePostIndex, parentmessageNamemessageargumentList.size());
-
-									messageargumentListInstance
-											.setRole(parentmessageNamemessageargumentList.get(k).getRole());
-									messageargumentListInstance
-											.setType(parentmessageNamemessageargumentList.get(k).getType());
-									String parentTag = parentmessageNamemessageargumentList.get(k).getTag();
-
-									ComplexType complexType = new ComplexType();
-
-									complexType.setMessageName(sourceMessageName);
-									for (int l = 0; l < messageargumentList.size(); l++) {
-										MessageargumentList messageargumentList12 = (MessageargumentList) messageargumentList
-												.get(l);
-										MessageargumentList complexMessageargumentInstance = new MessageargumentList();
-										complexMessageargumentInstance.setRole(messageargumentList12.getRole());
-										complexMessageargumentInstance.setType(messageargumentList12.getType());
-										complexMessageargumentInstance.setName(messageargumentList12.getName());
-										complexMessageargumentInstance
-												.setTag(parentTag + "." + messageargumentList12.getTag());
-										comlpexmessageargumentList.add(complexMessageargumentInstance);
-
-									}
-									complexType.setMessageargumentList(comlpexmessageargumentList);
-									messageargumentListInstance.setComplexType(complexType);
-									messageargumentListInstance
-											.setName(parentmessageNamemessageargumentList.get(k).getName());
-									messageargumentListInstance
-											.setTag(parentmessageNamemessageargumentList.get(k).getTag());
-									if (messageargumentListInstanceList != null
-											&& messageargumentListInstanceList.size() != 0) {
-										messageargumentListInstanceList.add(messageargumentListInstance);
-										expendedmessageBody.setMessageargumentList(messageargumentListInstanceList);
-									} else {
-										messageargumentListInstanceList
-												.addAll(messageargumentListInstanceSympletypePreList);
-										messageargumentListInstanceList.add(messageargumentListInstance);
-										messageargumentListInstanceList
-												.addAll(messageargumentListInstanceSympletypePostList);
-										expendedmessageBody.setMessageargumentList(messageargumentListInstanceList);
-									}
-									expendedChildmessageBodyList.add(sourceMessageBody.get(i).getMessageName());
-									if (message == destinationMessageBody.get(j).getMessageName()) {
-
-									} else {
-										expendedmessageBodyList.add(expendedmessageBody);
-									}
-
-								}
-								break;
-							} else {
-								isSubMessageFound = false;
-							}
-						}
-					}
-					if (isSubMessageFound) {
-						isSubMessageFound = true;
+			for (OutputMessage outputmsg : outputMessages) {
+				outputMessageName = outputmsg.getOutPutMessageName();
+				for (InputMessage inputmsg : inputMessages) {
+					if (inputmsg.getInputMessageName().equals(outputMessageName)) {
+						msgExist = true;
 						break;
 					}
 				}
-				if (!isSubMessageFound) {
-					parentNotExistList.add(sourceMessageBody.get(i));
-				}
-			}
-			// System.out.println(expendedmessageBodyList.size());
-			final int size = parentNotExistList.size();
-			boolean dulicateMessageIsFound = false;
-			for (int ij = 0; ij < size; ij++) {
-				String duplicateMessage = parentNotExistList.get(ij).getMessageName();
-				for (int j = 0; j < expendedmessageBodyList.size(); j++) {
-					if (duplicateMessage.equals(expendedmessageBodyList.get(j).getMessageName())) {
-						dulicateMessageIsFound = true;
-						break;
-					} else {
-						dulicateMessageIsFound = false;
+				if (!msgExist) {
+					messageBody = constructExpandedMessageBody(outputMessageName, protoBufClass.getListOfMessages(),
+							"");
+					if (null != messageBody) {
+						listOfMessages.add(messageBody);
 					}
 				}
-				if (!dulicateMessageIsFound) {
-					expendedmessageBodyList.add(parentNotExistList.get(ij));
-					dulicateMessageIsFound = false;
-				}
+				msgExist = false;
 			}
-		} catch (Exception ex) {
-			logger.error("-------------- constructconstructSubMessageBody() end ---------------" + ex);
-			ex.printStackTrace();
+
 		}
-		logger.debug("-------------- constructconstructSubMessageBody() end ---------------");
-		return expendedmessageBodyList;
+
+		return listOfMessages;
 	}
 
+	private MessageBody constructExpandedMessageBody(String messageName, List<MessageBody> listOfMessages, String tag) {
+		MessageBody result = null;
+
+		String basicProtobufTypes = Properties.getProtobufBasicType();
+		String[] basicProtobufTypesArray = basicProtobufTypes.split(",");
+		List<String> basicProtobufTypesList = Arrays.asList(basicProtobufTypesArray);
+
+		String parentTag = "";
+		if (null != tag && !tag.trim().equals("")) {
+			parentTag = tag + ".";
+		}
+		parentTag = parentTag.trim();
+		// get MessageBody for messageName
+		MessageBody sourceMessage = null;
+		for (MessageBody msgBody : listOfMessages) {
+			if (msgBody.getMessageName().equals(messageName)) {
+				sourceMessage = msgBody;
+				break;
+			}
+		}
+		if (null != sourceMessage) {
+			result = new MessageBody();
+			result.setMessageName(messageName);
+			List<MessageargumentList> messageargumentList = new ArrayList<MessageargumentList>();
+			MessageargumentList msgArgument = null;
+			String sourceMsgArgumentType = null;
+			List<MessageargumentList> sourceMessageargumentList = sourceMessage.getMessageargumentList();
+			ComplexType complexType = null;
+
+			for (MessageargumentList sourceMsgArgument : sourceMessageargumentList) {
+				sourceMsgArgumentType = sourceMsgArgument.getType();
+				msgArgument = new MessageargumentList();
+				msgArgument.setRole(sourceMsgArgument.getRole());
+				msgArgument.setType(sourceMsgArgumentType);
+				msgArgument.setName(sourceMsgArgument.getName());
+				msgArgument.setTag(parentTag + sourceMsgArgument.getTag());
+				// check source message argument is not basic type then
+				if (!basicProtobufTypesList.contains(sourceMsgArgumentType)) {
+					complexType = new ComplexType();
+					MessageBody complexBody = constructExpandedMessageBody(sourceMsgArgumentType, listOfMessages,
+							parentTag + sourceMsgArgument.getTag());
+					if (null != complexBody) {
+						complexType.setMessageName(complexBody.getMessageName());
+						complexType.setMessageargumentList(complexBody.getMessageargumentList());
+					}
+					msgArgument.setComplexType(complexType);
+				}
+				messageargumentList.add(msgArgument);
+			}
+			result.setMessageargumentList(messageargumentList);
+		}
+		return result;
+	}
 }